Change delegation of changed text
It now uses target/action to delegate text changes instead of teh method should change text. The method ShouldChangeText has been reduced to return the result of the respective delegate method. The check if the new text is the current text has been removed and must be moved to the app that uses SHSearchBar